resolved infinite loop issue
Looks like the base href value was a culprit. Once I replaced the
absolute value (document.location) with '/', the problem was resolved.
The other issue was '/' prefix in front of the templateURL values.
Fixing either resolves the infinite loop issue, but both need to be
fixed for proper and sturdier routing.